Domain specific routing
https://github.com/isaachinman/next-i18next/issues/16#issuecomment-520853281
