+++ public/static/locales/en/common.json
@@ -1,3 +1,4 @@
 {
+  "home": "Home",
   "helloWorld": "Hello World"
 }
