+++ public/static/locales/es/common.json
@@ -1,3 +1,4 @@
 {
+  "home": "Inicio",
   "helloWorld": "Hola mundo"
 }
