+++ src/pages/_app.js
@@ -1,5 +1,4 @@
-import { I18nextProvider } from "react-i18next";
-import i18n, { appWithTranslation } from "../i18n";
+import { appWithTranslation } from "../i18n";
 import { useRouter } from "next/router";
 import Layout from "../components/Layout";
 import "../styles.css";
@@ -7,8 +6,6 @@ import App from 'next/app'
 
 function MyApp({ Component, pageProps }) {
   const router = useRouter();
-  const { locale } = router.query;
-  console.log("App locale:", locale);
   return (
     <Layout>
       <Component {...pageProps} />
